Moved pages into components, convert to jsx
